+++ app/models/impact_report/metric.rb
@@ -67,7 +67,7 @@ module ImpactReport
       targets
     ].freeze
 
-    POTENTIAL_SAVINGS_METRICS = SchoolGroups::ImpactReport::PotentialSavings::METRICS
+    POTENTIAL_SAVINGS_METRICS = SchoolGroups::ImpactReport::Generator::PotentialSavings::METRICS
 
     FOOTNOTE_METRICS = %i[].freeze
 
@@ -79,9 +79,5 @@ module ImpactReport
       FOOTNOTE_METRICS).freeze
 
     enum :metric_type, enum_map(METRIC_TYPES).freeze, prefix: :type
-
-    def self.metrics(category)
-      const_get("#{category.upcase}_METRICS")
-    end
   end
 end
