+++ spec/services/school_groups/impact_report/generator/potential_savings_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe SchoolGroups::ImpactReport::PotentialSavings do
+RSpec.describe SchoolGroups::ImpactReport::Generator::PotentialSavings do
   subject(:potential_savings) { described_class.new(SchoolGroups::ImpactReport.new(school_group)) }
 
   let(:school_group) { create(:school_group) }
